LayoutTests/media/media-can-play-ogg-expected.txt

Test HTMLMediaElement canPlayType() method for ogg media containers.

EXPECTED (video.canPlayType('audio/ogg') == 'maybe') OK
EXPECTED (video.canPlayType('video/ogg') == 'maybe') OK
END OF TEST
